refactor: [M3-6402] - SRC > Components > TableContentWrapper
