--- packages/manager/src/components/TableContentWrapper/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './TableContentWrapper';
